django-template/requirements.txt
asgiref==3.8.1
Django==5.0.4
sqlparse==0.5.0
tzdata==2024.1
pillow
psycopg2-binary
djangorestframework
djangorestframework-simplejwt
django-cors-headers
python-dotenv
dj-database-url